+++ eea/userseditor/permissions.py
@@ -1,0 +1,1 @@
+EIONET_EDIT_USERS = 'Eionet edit users'